Refactor doGet method in ExchangeServlet class

+++ src/main/java/aslmk/Models/Exchange.java
@@ -1,4 +1,6 @@
 package aslmk.Models;
 
+import java.math.BigDecimal;
+
 public record Exchange(Currency baseCurrency, Currency targetCurrency,
-        double rate, double amount, double convertedAmount) {}
+                       double rate, BigDecimal amount, BigDecimal convertedAmount) {}
